Added styling and added components and pages
