+++ js-frameworks-ca/package.json
@@ -8,13 +8,16 @@
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
     "bootstrap": "^5.3.3",
+    "bootstrap-icons": "^1.11.3",
     "mdb-react-ui-kit": "^7.2.0",
     "mdb-ui-kit": "^7.2.0",
     "react": "^18.2.0",
     "react-bootstrap": "^2.10.1",
     "react-bootstrap-validation": "^0.1.11",
     "react-dom": "^18.2.0",
+    "react-fontawesome": "^1.7.1",
     "react-hook-form": "^7.51.0",
+    "react-router-bootstrap": "^0.26.2",
     "react-router-dom": "^6.22.2",
     "react-scripts": "5.0.1",
     "sass": "^1.71.1",
