+++ js-frameworks-ca/src/App.jsx
@@ -1,189 +1,17 @@
-import Button from 'react-bootstrap/Button';
-import Container from 'react-bootstrap/Container';
-import Form from 'react-bootstrap/Form';
-import Nav from 'react-bootstrap/Nav';
-import Navbar from 'react-bootstrap/Navbar';
 import { Routes, Route, Outlet } from "react-router-dom";
-import { MDBFooter } from 'mdb-react-ui-kit';
-import { useForm } from 'react-hook-form';
-import { yupResolver } from '@hookform/resolvers/yup';
-import * as yup from 'yup';
-import React, { useEffect, useState } from 'react';
+import React from 'react';
+import Contact from "./pages/Contact";
+import Footer from "./components/Footer";
+import Header from "./components/Header";
+import Products from "./pages/Products";
+
 
 function Home() {
   return <div className='vh-100'>Home</div>
 }
 
-function Products() {
-
-  const url = 'https://v2.api.noroff.dev/online-shop';
-
-    const [posts, setPosts] = useState([]);
-    // State for holding our loading state
-    const [isLoading, setIsLoading] = useState(false);
-    // State for holding our error state
-    const [isError, setIsError] = useState(false);
-  
-    useEffect(() => {
-      async function getData() {
-        try {
-          // Reset the error state in case there as an error previously
-          setIsError(false);
-          // Turn on the loading state each time we do an API call
-          setIsLoading(true);
-          const response = await fetch(url);
-          const json = await response.json();
-          console.log(json.data);
-          console.log(response);
-          setPosts(json.data);
-          // Clear the loading state once we've successfully got our data
-          setIsLoading(false);
-        } catch (error) {
-          // Clear the loading state if we get an error and then
-          // set our error state to true
-          setIsLoading(false);
-          setIsError(true);
-        }
-      }
-  
-      getData();
-    }, []);
-  
-    if (isLoading) {
-      return <div>Loading posts</div>;
-    }
-  
-    if (isError) {
-      return <div>Error loading data</div>;
-    }
-  
-    return (
-      <div className='mx-5'> <h1 className="text-center p-3">Products</h1>
-       {Array.from(posts).map((post) => {
-          return (
-            <div className='border p-3 mt-2 text-center w-50'>
-            <div key={post.id}>
-              <h2 className='text-center'>{post.title}</h2>
-              <img src={post.image.url} className='w-75'></img>
-              <p>{post.description}</p>
-              {/* <img src={post.image.url}></img> */}
-              <p>{post.discountedPrice}$</p>
-              <p>Rating: {post.rating}</p>
-              {/* <h6>Reviews: {post.reviews[0].description}</h6> */}
-            </div>
-            </div>
-          );
-        })}
-      </div>
-    );
-  }
-
-function Contact() {
-  // return <div>Contact</div>
-  const schema = yup
-  .object({
-    fullName: yup
-      .string()
-      .min(3, 'Your full name should be at least 3 characters.')
-      .max(50, 'Your full name cannot be longer than 50 characters.')
-      .required('Please enter your full name'),
-    email: yup
-      .string().email()
-      .required('Your email must be @noroff.no'),
-    subject: yup
-      .string()
-      .min(3, 'Your subject should be at least 3 characters.')
-      .max(100, 'Your subject cannot be longer than 1000 characters.')
-      .required('Please enter your subject'),
-    body: yup
-      .string()
-      .min(3, 'Your body should be at least 3 characters.')
-      .max(1000, 'Your body cannot be longer than 10 characters.')
-      .required('Please enter your body'),
-  })
-  .required();
-
-  const {
-        register,
-        handleSubmit,
-        formState: { errors },
-      } = useForm({
-        resolver: yupResolver(schema),
-      });
-    
-      function onSubmit(data) {
-        console.log(data);
-      }
-    
-      return (
-        <form onSubmit={handleSubmit(onSubmit)} className='vh-100'>
-          <input {...register('fullName')} />
-          <p>{errors.fullName?.message}</p>
-          <input {...register('email')} />
-          <p>{errors.email?.message}</p>
-          <input {...register('subject')} />
-          <p>{errors.subject?.message}</p>
-          <input {...register('body')} />
-          <p>{errors.body?.message}</p>
-          <input type="submit" />
-        </form>
-      );
-}
-
 function RouteNotFound() {
   return <div>Page not found</div>
-}
-
-function NavHeader() {
-  return (
-    <Navbar expand="md" className="bg-body-tertiary">
-      <Container fluid>
-        <Navbar.Brand href="/">ECommerce Store</Navbar.Brand>
-        <Navbar.Toggle aria-controls="navbarScroll" />
-        <Navbar.Collapse id="navbarScroll">
-          <Nav
-            className="me-auto my-2 my-lg-0"
-            style={{ maxHeight: '100px' }}
-            navbarScroll
-          >
-            <Nav.Link href="/">Home</Nav.Link>
-            <Nav.Link href="/products">Products</Nav.Link>
-            <Nav.Link href="/contact">Contact</Nav.Link>
-          </Nav>
-          <Form className="d-flex">
-            <Form.Control
-              type="search"
-              placeholder="Search"
-              className="me-2"
-              aria-label="Search"
-            />
-            <Button variant="outline-success">Search</Button>
-          </Form>
-        </Navbar.Collapse>
-      </Container>
-    </Navbar>
-  );
-}
-
-// Our header component that gets used in our <Layout> component
-function Header() {
-  return (
-    <header>
-      <NavHeader />
-    </header>
-  )
-}
-
-// Our footer component that gets used in our <Layout> component
-function Footer() {
-
-  return (
-    <MDBFooter className='text-center text-lg-left bottom-0 w-100 bg-body-tertiary'>
-    <div className='text-center p-3' >
-      &copy; {new Date().getFullYear()} ECommerce Store{' '}
-    </div>
-  </MDBFooter>
-  );
 }
 
 function Layout() {
