+++ js-frameworks-ca/src/components/Header/index.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import NavBar from '../NavBar';
+
+// Our header component that gets used in our <Layout> component
+function Header() {
+    return (
+      <header>
+        <NavBar />
+      </header>
+    )
+  }
+
+  export default Header;
