+++ js-frameworks-ca/src/components/NavBar/index.jsx
@@ -1,0 +1,45 @@
+import Button from 'react-bootstrap/Button';
+import Container from 'react-bootstrap/Container';
+import Form from 'react-bootstrap/Form';
+import Nav from 'react-bootstrap/Nav';
+import Navbar from 'react-bootstrap/Navbar';
+import React from 'react';
+import { NavLink } from 'react-router-dom';
+
+
+function NavBar() {
+    return (
+      <Navbar expand="md" className="bg-body-tertiary">
+        <Container fluid>
+          <Navbar.Brand as={NavLink} to="/">ECommerce Store</Navbar.Brand>
+          <Navbar.Toggle aria-controls="navbarScroll" />
+          <Navbar.Collapse id="navbarScroll">
+            <Nav
+              className="me-auto my-2 my-lg-0"
+              style={{ maxHeight: '100px' }}
+              navbarScroll
+            >
+              {/* <Nav.Link as={NavLink} to="/">Home</Nav.Link> */}
+              <Nav.Link as={NavLink} to="/products">Products</Nav.Link>
+              <Nav.Link as={NavLink} to="/contact">Contact</Nav.Link>
+              <Nav.Link as={NavLink} to="/cart">
+                <i class="bi bi-cart-fill text-dark me-1"></i>
+                0
+              </Nav.Link>
+            </Nav>
+            <Form className="d-flex">
+              <Form.Control
+                type="search"
+                placeholder="Search"
+                className="me-0"
+                aria-label="Search"
+              />
+              <Button variant="outline-dark">Search</Button>
+            </Form>
+          </Navbar.Collapse>
+        </Container>
+      </Navbar>
+    );
+  }
+
+  export default NavBar;
