+++ js-frameworks-ca/src/index.js
@@ -5,7 +5,7 @@ import App from './App';
 import reportWebVitals from './reportWebVitals';
 import '../node_modules/bootstrap/dist/css/bootstrap.min.css';
 import { BrowserRouter } from 'react-router-dom';
-
+import "bootstrap-icons/font/bootstrap-icons.css";
 // ...imports above
 
 const root = ReactDOM.createRoot(document.getElementById("root"))
